+++ core-ng/src/main/java/core/framework/impl/web/response/FileBodyResponseHandler.java
@@ -16,22 +16,23 @@ import java.nio.channels.FileChannel;
 /**
  * @author neo
  */
-public class FileBodyResponseHandler implements BodyHandler {
+class FileBodyResponseHandler implements BodyHandler {
     @Override
     public void handle(ResponseImpl response, Sender sender, RequestImpl request) {
         File file = ((FileBody) response.body).file;
-
         try {
             final FileChannel channel = new FileInputStream(file).getChannel();
             sender.transferFrom(channel, new IoCallback() {
                 @Override
                 public void onComplete(HttpServerExchange exchange, Sender sender) {
                     IoUtils.safeClose(channel);
+                    END_EXCHANGE.onComplete(exchange, sender);
                 }
 
                 @Override
                 public void onException(HttpServerExchange exchange, Sender sender, IOException exception) {
                     IoUtils.safeClose(channel);
+                    END_EXCHANGE.onException(exchange, sender, exception);
                     throw new UncheckedIOException(exception);
                 }
             });